feat(read-emails): consolidate node examples
