+++ packages/read-emails/backend/node/server.js
@@ -1,16 +1,19 @@
+const express = require('express');
+const cors = require('cors');
 const dotenv = require('dotenv');
-const url = require('url');
 const mockDb = require('./utils/mock-db');
-const { mockServer, getReqBody } = require('./utils/mock-server');
 
 const Nylas = require('nylas');
 const { WebhookTriggers } = require('nylas/lib/models/webhook');
 const { Scope } = require('nylas/lib/models/connect');
-const { Routes: NylasRoutes } = require('nylas/lib/services/routes');
-const { DefaultPaths } = require('nylas/lib/services/routes');
 const { openWebhookTunnel } = require('nylas/lib/services/tunnel');
 
 dotenv.config();
+
+const app = express();
+
+// Enable CORS
+app.use(cors());
 
 // The port the express app will run on
 const port = 9000;
@@ -25,109 +28,44 @@ const nylasClient = new Nylas({
 const CLIENT_URI =
   process.env.CLIENT_URI || `http://localhost:${process.env.PORT || 3000}`;
 
-// Use the routes provided by the Nylas Node SDK to quickly implement
-// the authentication flow
-const { buildAuthUrl, exchangeCodeForToken } = NylasRoutes(nylasClient);
+// '/nylas/generate-auth-url': This route builds the URL for
+// authenticating users to your Nylas application via Hosted Authentication
+app.post('/nylas/generate-auth-url', express.json(), async (req, res) => {
+  const { body } = req;
 
-// Configure the Nylas routes using your flavour of backend framework
-// '/nylas/generate-auth-url': This route builds the URL for authenticating
-// users to your Nylas application via Hosted Authentication
-mockServer.post(DefaultPaths.buildAuthUrl, async (req, res) => {
-  const body = await getReqBody(req);
-
-  const authUrl = await buildAuthUrl({
+  const authUrl = nylasClient.urlForAuthentication({
+    loginHint: body.email_address,
+    redirectURI: (CLIENT_URI || '') + body.success_url,
     scopes: [Scope.EmailReadOnly],
-    emailAddress: body.email_address,
-    successUrl: body.success_url,
-    clientUri: CLIENT_URI,
   });
 
-  res.writeHead(200).end(authUrl);
+  return res.send(authUrl);
 });
 
-// '/nylas/exchange-mailbox-token': This route exchanges an
-// authorization code for an access token
-mockServer.post(DefaultPaths.exchangeCodeForToken, async (req, res) => {
-  const body = await getReqBody(req);
+// '/nylas/exchange-mailbox-token': This route exchanges an authorization
+// code for an access token
+// and sends the details of the authenticated user to the client
+app.post('/nylas/exchange-mailbox-token', express.json(), async (req, res) => {
+  const body = req.body;
 
-  try {
-    const { accessToken, emailAddress } = await exchangeCodeForToken(
-      body.token
-    );
+  const { accessToken, emailAddress } = await nylasClient.exchangeCodeForToken(
+    body.token
+  );
 
-    // Normally store the access token in the DB
-    console.log('Access Token was generated for: ' + emailAddress);
-    // Replace this mock code with your actual database operations
-    const user = await mockDb.createOrUpdateUser(emailAddress, {
-      accessToken,
-      emailAddress,
-    });
+  // Normally store the access token in the DB
+  console.log('Access Token was generated for: ' + emailAddress);
 
-    // Return an authorization object to the user
-    res.writeHead(200).end(
-      JSON.stringify({
-        id: user.id,
-        emailAddress: user.emailAddress,
-      })
-    );
-  } catch (e) {
-    res.writeHead(500).end(e.message);
-  }
-});
+  // Replace this mock code with your actual database operations
+  const user = await mockDb.createOrUpdateUser(emailAddress, {
+    accessToken,
+    emailAddress,
+  });
 
-// Add route for getting 5 latest emails
-mockServer.get('/nylas/read-emails', async (req, res) => {
-  if (!req.headers.authorization) {
-    return res.writeHead(401).end('Unauthorized');
-  }
-
-  const user = await mockDb.findUser(req.headers.authorization);
-  if (!user) {
-    return res.writeHead(401).end('Unauthorized');
-  }
-
-  const threads = await nylasClient
-    .with(user.accessToken)
-    .threads.list({ limit: 5, expanded: true });
-
-  return res.writeHead(200).end(JSON.stringify(threads));
-});
-
-// Add route for fetching message
-mockServer.get('/nylas/message', async (req, res) => {
-  if (!req.headers.authorization) {
-    return res.writeHead(401).end('Unauthorized');
-  }
-
-  const user = await mockDb.findUser(req.headers.authorization);
-  if (!user) {
-    return res.writeHead(401).end('Unauthorized');
-  }
-
-  const { id } = url.parse(req.url, true).query;
-  const message = await nylasClient.with(user.accessToken).messages.find(id);
-
-  return res.writeHead(200).end(JSON.stringify(message));
-});
-
-// Add route for download file
-mockServer.get('/nylas/file', async (req, res) => {
-  if (!req.headers.authorization) {
-    return res.writeHead(401).end('Unauthorized');
-  }
-
-  const user = await mockDb.findUser(req.headers.authorization);
-  if (!user) {
-    return res.writeHead(401).end('Unauthorized');
-  }
-
-  const { id } = url.parse(req.url, true).query;
-  const file = await nylasClient.with(user.accessToken).files.find(id);
-
-  // Files will be returned as a binary object
-  const fileData = await file.download();
-
-  return res.writeHead(200).end(fileData?.body);
+  // Return an authorization object to the user
+  return res.json({
+    id: user.id,
+    emailAddress: user.emailAddress,
+  });
 });
 
 // Start the Nylas webhook
@@ -143,12 +81,64 @@ openWebhookTunnel(nylasClient, {
         break;
     }
   },
-}).then((webhookDetails) =>
-  console.log('Webhook tunnel registered. Webhook ID: ' + webhookDetails.id)
-);
+}).then((webhookDetails) => {
+  console.log('Webhook tunnel registered. Webhook ID: ' + webhookDetails.id);
+});
 
-// Before we start our backend, we should whitelist our frontend as a
-// redirect URI to ensure the auth completes
+// Middleware to check if the user is authenticated
+async function isAuthenticated(req, res, next) {
+  if (!req.headers.authorization) {
+    return res.status(401).json('Unauthorized');
+  }
+
+  // Query our mock db to retrieve the stored user access token
+  const user = await mockDb.findUser(req.headers.authorization);
+
+  if (!user) {
+    return res.status(401).json('Unauthorized');
+  }
+
+  // Add the user to the response locals
+  res.locals.user = user;
+
+  next();
+}
+
+// Add route for getting 5 latest emails
+app.get('/nylas/read-emails', isAuthenticated, async (req, res) => {
+  const user = res.locals.user;
+
+  const threads = await nylasClient
+    .with(user.accessToken)
+    .threads.list({ limit: 5, expanded: true });
+
+  return res.json(threads);
+});
+
+// Add route for getting individual message by id
+app.get('/nylas/message', isAuthenticated, async (req, res) => {
+  const user = res.locals.user;
+
+  const { id } = req.query;
+  const message = await nylasClient.with(user.accessToken).messages.find(id);
+
+  return res.json(message);
+});
+
+// Add route for downloading file
+app.get('/nylas/file', isAuthenticated, async (req, res) => {
+  const user = res.locals.user;
+
+  const { id } = req.query;
+  const file = await nylasClient.with(user.accessToken).files.find(id);
+
+  // Files will be returned as a binary object
+  const fileData = await file.download();
+  return res.end(fileData?.body);
+});
+
+// Before we start our backend, we should whitelist our frontend
+// as a redirect URI to ensure the auth completes
 nylasClient
   .application({
     redirectUris: [CLIENT_URI],
@@ -156,10 +146,9 @@ nylasClient
   .then((applicationDetails) => {
     console.log(
       'Application whitelisted. Application Details: ',
-      JSON.stringify(applicationDetails, undefined, 2)
+      JSON.stringify(applicationDetails)
     );
   });
 
 // Start listening on port 9000
-mockServer.init().listen(port);
-console.log('App listening on port ' + port);
+app.listen(port, () => console.log('App listening on port ' + port));
